Manage service information in Sanity
https://app.clickup.com/t/86bwxn47y

The copy for services that we include in the proposal and pricing menus
will now be managed in Sanity. The details for the website remains.

Includes a lot of validation rules to try to prevent any extra or
missing information based on which destinations have been included.

The website queries for the services have also been updated to
accommodate these changes.