Fix bug with catch-up fee calculation
Agreements where there was a self assessment for a previous year and
nothing for the future would result in a single payment being taken
because of an incomplete calculation to spread the cost across the
remaining agreement but no subscription was started.

Now the entire catch-up fee is taken and no subscription is started.

Also added some tests, using Claude, to try to avoid adding new
bugs.